requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asttokens=2.0.8=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=2.116=mkl
blas-devel=3.9.0=16_win64_mkl
brotlipy=0.7.0=py310he2412df_1004
bzip2=1.0.8=h8ffe710_4
ca-certificates=2022.9.24=h5b45459_0
certifi=2022.9.24=pyhd8ed1ab_0
cffi=1.15.1=py310hcbf9ad4_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
colorama=0.4.5=pyhd8ed1ab_0
cryptography=37.0.1=py310h21b164f_0
cudatoolkit=11.6.0=hc0ea762_10
debugpy=1.6.3=py310h8a704f9_0
decorator=5.1.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
executing=1.1.1=pyhd8ed1ab_0
freetype=2.12.1=h546665d_0
idna=3.4=pyhd8ed1ab_0
intel-openmp=2022.1.0=h57928b3_3787
ipykernel=6.16.0=pyh025b116_0
ipython=8.5.0=pyh08f2357_1
jedi=0.18.1=pyhd8ed1ab_2
joblib=1.2.0=pyhd8ed1ab_0
jpeg=9e=h8ffe710_2
jupyter_client=7.4.2=pyhd8ed1ab_0
jupyter_core=4.11.1=py310h5588dad_0
lcms2=2.12=h2a16943_0
lerc=4.0.0=h63175ca_0
libblas=3.9.0=16_win64_mkl
libcblas=3.9.0=16_win64_mkl
libdeflate=1.14=hcfcfb64_0
libffi=3.4.2=h8ffe710_5
liblapack=3.9.0=16_win64_mkl
liblapacke=3.9.0=16_win64_mkl
libpng=1.6.38=h19919ed_0
libsodium=1.0.18=h8d14728_1
libsqlite=3.39.4=hcfcfb64_0
libtiff=4.4.0=h8e97e67_4
libuv=1.44.2=h8ffe710_0
libwebp-base=1.2.4=h8ffe710_0
libxcb=1.13=hcd874cb_1004
libzlib=1.2.13=hcfcfb64_4
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2022.1.0=h6a75c08_874
mkl-devel=2022.1.0=h57928b3_875
mkl-include=2022.1.0=h6a75c08_874
msys2-conda-epoch=20160418=1
nest-asyncio=1.5.6=pyhd8ed1ab_0
numpy=1.23.4=py310h4a8f9c9_0
openjpeg=2.5.0=hc9384bd_1
openssl=3.0.5=hcfcfb64_2
packaging=21.3=pyhd8ed1ab_0
pandas=1.5.0=py310h1c4a608_0
parso=0.8.3=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=9.2.0=py310h52929f7_2
pip=22.3=pyhd8ed1ab_0
prompt-toolkit=3.0.31=pyha770c72_0
psutil=5.9.2=py310h8d17308_0
pthread-stubs=0.4=hcd874cb_1001
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.13.0=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_1
pyparsing=3.0.9=pyhd8ed1ab_0
pysocks=1.7.1=pyh0701188_6
python=3.10.6=hcf16a7b_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.10=2_cp310
pytorch=1.12.1=py3.10_cuda11.6_cudnn8_0
pytorch-mutex=1.0=cuda
pytz=2022.4=pyhd8ed1ab_0
pywin32=303=py310he2412df_0
pyzmq=24.0.1=py310hcd737a0_0
requests=2.28.1=pyhd8ed1ab_1
scikit-learn=1.1.2=py310h3a564e9_0
scipy=1.9.1=py310h578b7cb_0
setuptools=65.5.0=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
stack_data=0.5.1=pyhd8ed1ab_0
tbb=2021.6.0=h91493d7_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=h8ffe710_0
torchaudio=0.12.1=py310_cu116
torchvision=0.13.1=py310_cu116
tornado=6.2=py310he2412df_0
traitlets=5.5.0=pyhd8ed1ab_0
typing_extensions=4.4.0=pyha770c72_0
tzdata=2022e=h191b570_0
ucrt=10.0.20348.0=h57928b3_0
urllib3=1.26.11=pyhd8ed1ab_0
vc=14.2=hac3ee0b_8
vs2015_runtime=14.29.30139=h890b9b1_8
wcwidth=0.2.5=pyh9f0ad1d_2
wheel=0.37.1=pyhd8ed1ab_0
win_inet_pton=1.1.0=py310h5588dad_4
xorg-libxau=1.0.9=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.2.6=h8d14728_0
zeromq=4.3.4=h0e60522_1
zstd=1.5.2=h7755175_4
